lavr.MSH-29.export (#42)
* MSH-27: add unset

* add export function

* MSH-29:modified export

* MSH-29:add status for function export

* MSH-29:clear check identifier function

* clear export function

* change heders files

* MSH-29: add payload in function export

* add i = 0

+++ errors/errors_factory.c
@@ -35,7 +35,7 @@ static const char *(*err_resolver(t_err_domain domain))(int)
         [ERR_EXEC] = err_exec_msg};
     const size_t size = sizeof table / sizeof *table;
 
-    if (domain < 0 || domain >= (int)size)
+    if (domain < 0 || domain >= size)
         return NULL;
     return table[domain];
 }
